Extract header into component
